+++ src/main/java/com/ime/lockmanager/user/application/port/in/res/ApplyingStudentsDto.java
@@ -6,7 +6,7 @@ import lombok.Getter;
 
 @Builder
 @Getter
-public class AllApplyingStudentDto {
+public class ApplyingStudentsDto {
     @Schema(defaultValue = "학생회비 납부 신청한 학생의 학번")
     private String studentNum;
     @Schema(defaultValue = "학생회비 납부 신청한 학생의 이름")
